+++ test/buffer_reader_test.dart
@@ -1,332 +1,192 @@
 import 'dart:math';
 import 'dart:typed_data';
 
-import 'package:convert/convert.dart';
 import 'package:dart_buffer/dart_buffer.dart';
 import 'package:test/test.dart';
 
+import 'common_helper.dart';
+
+void testValue(BufferReader bufferReader, dynamic value, dynamic expectedValue,
+    [int? expectedOffset]) {
+  expectedOffset ??= expectedValue is List<int> ? expectedValue.length : 0;
+
+  expect(bufferReader.offset, expectedOffset);
+
+  if (expectedValue is List<int> && value is List<int>) {
+    expect(
+      value.sublist(
+          0, value.length >= expectedOffset ? expectedOffset : value.length),
+      expectedValue.sublist(
+          0,
+          expectedValue.length >= expectedOffset
+              ? expectedOffset
+              : expectedValue.length),
+    );
+  } else {
+    expect(value as int, expectedValue);
+  }
+}
+
 void main() {
-  final numbers = List.generate(16, (index) => Random().nextInt(10).abs());
-  final numbersNegative = List.generate(16, (index) {
-    int num = 0;
-    do {
-      num = Random().nextInt(10).abs() * -1;
-    } while (num == 0);
+  test('BufferReader.getInt8()', () {
+    final values = [-128, -127, -1, 0, 1, 0x7d, 0x7f];
+    final buffer = Int8List.fromList(values).buffer.asByteData();
+    final bufferReader = BufferReader(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 1;
+      final val = bufferReader.getInt8();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getUInt8()', () {
+    final values = [0, 1, 0xfe, 0xff];
+    final buffer = Uint8List.fromList(values).buffer.asByteData();
+    final bufferReader = BufferReader(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 1;
+      final val = bufferReader.getUInt8();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getInt16()', () {
+    final values = [-32768, -32767, -1, 0, 1, 0x7ffd, 0x7fff];
+    final buffer = Int16List.fromList(values).buffer.asByteData();
+    final bufferReader = BufferReader(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 2;
+      final val = bufferReader.getInt16();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getUInt16()', () {
+    final values = [0, 1, 0x7ffd, 0x7fff];
+    final buffer = Uint16List.fromList(values).buffer.asByteData();
+    final bufferReader = BufferReader(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 2;
+      final val = bufferReader.getUInt16();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getInt32()', () {
+    final maxInt32 = 2147483647;
+    final minInt32 = -2147483648;
+    final values = <int>[
+      minInt32,
+      minInt32 ~/ 2,
+      -1,
+      0,
+      1,
+      maxInt32 ~/ 2,
+      maxInt32
+    ];
+    final buffer = Int32List.fromList(values);
+    final bufferReader = BufferReader.fromTypedData(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 4;
+      final val = bufferReader.getInt32();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getUInt32()', () {
+    final maxInt32 = 2147483647;
+    final values = <int>[0, 1, maxInt32 ~/ 2, maxInt32];
+    final buffer = Uint32List.fromList(values);
+    final bufferReader = BufferReader.fromTypedData(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 4;
+      final val = bufferReader.getUInt32();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getInt64()', () {
+    final maxInt64 = 9223372036854775807;
+    final minInt64 = -9223372036854775808;
+    final values = <int>[
+      minInt64,
+      minInt64 ~/ 2,
+      -1,
+      0,
+      1,
+      maxInt64 ~/ 2,
+      maxInt64
+    ];
+    final buffer = Int64List.fromList(values);
+    final bufferReader = BufferReader.fromTypedData(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 8;
+      final val = bufferReader.getInt64();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getUInt64()', () {
+    final maxInt64 = 9223372036854775807;
+    final values = <int>[0, 1, maxInt64 ~/ 2, maxInt64];
+    final buffer = Uint64List.fromList(values);
+    final bufferReader = BufferReader.fromTypedData(buffer);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + 8;
+      final val = bufferReader.getUInt64();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getVarInt()', () {
+    final hexStr =
+        '0001fcfdfd00fdfe00fdff00fd0001fdfefffdfffffe00000100fefefffffffeffffffffff0000000001000000ffffffffffffff1f00';
+    final values = [
+      0,
+      1,
+      252,
+      253,
+      254,
+      255,
+      256,
+      pow(2, 16).toInt() - 2,
+      pow(2, 16).toInt() - 1,
+      pow(2, 16).toInt(),
+      pow(2, 32).toInt() - 2,
+      pow(2, 32).toInt() - 1,
+      pow(2, 32).toInt(),
+      9007199254740991,
+    ];
+    final bufferReader = BufferReader.fromHex(hexStr);
+    for (var v in values) {
+      int expectedOffset = bufferReader.offset;
+      final val = bufferReader.getVarInt();
+      expectedOffset += encodingLength(val);
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getSlice()', () {
+    final hexStr = '0101020304feff';
+    final values = [
+      [1],
+      [1, 2, 3, 4],
+      [254, 255]
+    ];
+    final bufferReader = BufferReader.fromHex(hexStr);
+    for (var v in values) {
+      final expectedOffset = bufferReader.offset + v.length;
+      final val = bufferReader.getSlice(v.length).buffer.asUint8List();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
+  });
+  test('BufferReader.getVarSlice()', () {
+    const hexStr =
+        '0101fc020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202020202fdfd0003030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303';
 
-    return num;
-  });
-  final buffPositive = Uint8List.fromList(numbers);
-  final buffNegative = Int8List.fromList(numbersNegative);
+    final values = [
+      Uint8List.fromList(List.filled(1, 1)),
+      Uint8List.fromList(List.filled(252, 2)),
+      Uint8List.fromList(List.filled(253, 3))
+    ];
 
-  group('constructors', () {
-    test('BufferReader.fromList(Uint8List)', () {
-      BufferReader? bufferReader;
-      try {
-        bufferReader = BufferReader.fromTypedData(buffPositive);
-      } catch (e) {
-        expect(e, isNull);
-      }
-
-      expect(bufferReader, isA<BufferReader>());
-    });
-    test('BufferReader.fromList(Int8List)', () {
-      BufferReader? bufferReader;
-      try {
-        bufferReader =
-            BufferReader.fromTypedData(buffPositive.buffer.asInt8List());
-      } catch (e) {
-        expect(e, isNull);
-      }
-
-      expect(bufferReader, isA<BufferReader>());
-    });
-    test('BufferReader()', () {
-      BufferReader? bufferReader;
-      try {
-        bufferReader = BufferReader(buffPositive.buffer.asByteData());
-      } catch (e) {
-        expect(e, isNull);
-      }
-
-      expect(bufferReader, isA<BufferReader>());
-    });
-  });
-  group('BufferReader.getInt8()', () {
-    test('should be work positive', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getInt8();
-      expect(actual, equals(buffPositive[0]));
-    });
-    test('should be work negative', () {
-      final bufferReader = BufferReader.fromTypedData(buffNegative);
-      final actual = bufferReader.getInt8();
-      expect(actual, equals(buffNegative[0]));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getInt8();
-      } catch (e) {
-        expect(e, isA<IndexError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getUInt8()', () {
-    test('should be work', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getUInt8();
-      expect(actual, equals(buffPositive[0]));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getUInt8();
-      } catch (e) {
-        expect(e, isA<IndexError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getInt16()', () {
-    test('should be work positive', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getInt16();
-      expect(
-          actual,
-          equals(int.parse(
-              hex.encode(buffPositive.sublist(0, 2).reversed.toList()),
-              radix: 16)));
-    });
-    test('should be work negative', () {
-      final bufferReader = BufferReader.fromTypedData(buffNegative);
-      final expected =
-          Int8List.fromList(buffNegative.sublist(0, 2).reversed.toList())
-              .buffer
-              .asByteData()
-              .getInt16(bufferReader.offset);
-      final actual = bufferReader.getInt16();
-      expect(actual, equals(expected));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getInt16();
-      } catch (e) {
-        expect(e, isA<RangeError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getUInt16()', () {
-    test('should be work', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getUInt16();
-      expect(
-          actual,
-          equals(int.parse(
-              hex.encode(buffPositive.sublist(0, 2).reversed.toList()),
-              radix: 16)));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getUInt16();
-      } catch (e) {
-        expect(e, isA<RangeError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getInt32()', () {
-    test('should be work positive', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getInt32();
-      expect(
-          actual,
-          equals(int.parse(
-              hex.encode(buffPositive.sublist(0, 4).reversed.toList()),
-              radix: 16)));
-    });
-    test('should be work negative', () {
-      final bufferReader = BufferReader.fromTypedData(buffNegative);
-      final expected =
-          Int8List.fromList(buffNegative.sublist(0, 4).reversed.toList())
-              .buffer
-              .asByteData()
-              .getInt32(bufferReader.offset);
-      final actual = bufferReader.getInt32();
-      expect(actual, equals(expected));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getInt32();
-      } catch (e) {
-        expect(e, isA<RangeError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getUInt32()', () {
-    test('should be work', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getUInt32();
-      expect(
-          actual,
-          equals(int.parse(
-              hex.encode(buffPositive.sublist(0, 4).reversed.toList()),
-              radix: 16)));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getUInt32();
-      } catch (e) {
-        expect(e, isA<RangeError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getInt64()', () {
-    test('should be work positive', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getInt64();
-      expect(
-          actual,
-          equals(int.parse(
-              hex.encode(buffPositive.sublist(0, 8).reversed.toList()),
-              radix: 16)));
-    });
-    test('should be work negative', () {
-      final bufferReader = BufferReader.fromTypedData(buffNegative);
-      final expected =
-          Int8List.fromList(buffNegative.sublist(0, 8).reversed.toList())
-              .buffer
-              .asByteData()
-              .getInt64(bufferReader.offset);
-      final actual = bufferReader.getInt64();
-      expect(actual, equals(expected));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getInt64();
-      } catch (e) {
-        expect(e, isA<RangeError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getUInt64()', () {
-    test('should be work', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getUInt64();
-      expect(
-          actual,
-          equals(int.parse(
-              hex.encode(buffPositive.sublist(0, 8).reversed.toList()),
-              radix: 16)));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getUInt64();
-      } catch (e) {
-        expect(e, isA<RangeError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getVarInt()', () {
-    test('8bit should be work', () {
-      final list = Uint8List.fromList([0xfc]);
-      final bufferReader = BufferReader.fromTypedData(list);
-      final actual = bufferReader.getVarInt();
-      expect(actual, equals(list[0]));
-    });
-    test('16bit should be work', () {
-      final list =
-          Uint8List.fromList([0xfd] + List.generate(2, (index) => index));
-      final bufferReader = BufferReader.fromTypedData(list);
-      final actual = bufferReader.getVarInt();
-      final value = int.parse(hex.encode(list.reversed.toList().sublist(0, 2)),
-          radix: 16);
-      expect(actual, equals(value));
-    });
-    test('32bit should be work', () {
-      final list =
-          Uint8List.fromList([0xfe] + List.generate(4, (index) => index));
-      final bufferReader = BufferReader.fromTypedData(list);
-      final actual = bufferReader.getVarInt();
-      final value = int.parse(hex.encode(list.reversed.toList().sublist(0, 4)),
-          radix: 16);
-      expect(actual, equals(value));
-    });
-    test('64bit should be work', () {
-      final list =
-          Uint8List.fromList([0xff] + List.generate(8, (index) => index));
-      final bufferReader = BufferReader.fromTypedData(list);
-      final actual = bufferReader.getVarInt();
-      final value = BigInt.parse(
-          hex.encode(list.reversed.toList().sublist(0, 8)),
-          radix: 16);
-      expect(actual, equals(value.toInt()));
-    });
-  });
-  group('BufferReader.getSlice()', () {
-    test('should be return ByteData', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      final actual = bufferReader.getSlice(10).buffer.asUint8List();
-      final expected = buffPositive.sublist(0, 10);
-      expect(hex.encode(actual), equals(hex.encode(expected)));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getSlice(10);
-      } catch (e) {
-        expect(e, isA<IndexError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
-  });
-  group('BufferReader.getVarSlice()', () {
-    test('should be return ByteData', () {
-      final bufferReader =
-          BufferReader.fromTypedData(Uint8List.fromList([10, ...buffPositive]));
-      final actual = bufferReader.getVarSlice().buffer.asUint8List();
-      final expected = buffPositive.sublist(0, 10);
-      expect(hex.encode(actual), equals(hex.encode(expected)));
-    });
-    test('should be throws error', () {
-      final bufferReader = BufferReader.fromTypedData(buffPositive);
-      bufferReader.offset = bufferReader.length;
-      try {
-        bufferReader.getVarSlice();
-      } catch (e) {
-        expect(e, isA<IndexError>());
-        return;
-      }
-      throw Exception('Should be throws.');
-    });
+    final bufferReader = BufferReader.fromHex(hexStr);
+    for (var v in values) {
+      final expectedOffset =
+          bufferReader.offset + encodingLength(v.length) + v.length;
+      final val = bufferReader.getVarSlice().buffer.asUint8List();
+      testValue(bufferReader, val, v, expectedOffset);
+    }
   });
 }