+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 2.0.0
+
+- add new methods
+- update convert methods
+
+## 1.0.3
+
+- fixed array bug
+
+## 1.0.2
+
+- offset increment bug fixed
+- update tests
+- add encodingLength function
+
 ## 1.0.1
 
 - add getUInt, setUInt method